CF2/requirements.txt
# Function dependencies, for example:
# package>=version
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
